+++ src/apps/settings/modules/information/infrastructure.tsx
@@ -1,26 +1,38 @@
 import { Icon } from '../../../shared/components/Icon';
 import { SettingsGroup, SettingsOption, SettingsSubGroup } from '../../components/SettingsBox';
-import { invoke } from '@tauri-apps/api/core';
 import { exit, relaunch } from '@tauri-apps/plugin-process';
-import { Button } from 'antd';
+import { Button, Switch } from 'antd';
+import { useDispatch, useSelector } from 'react-redux';
 
-import { EnvConfig, getSettingsPath } from '../shared/config/infra';
-import { LoadCustomConfigFile } from './infra.api';
+import { EnvConfig } from '../shared/config/infra';
 import cs from './infra.module.css';
 
+import { newSelectors, RootActions } from '../shared/store/app/reducer';
+
 export function Information() {
-  const openSettingsFile = async () => {
-    invoke('open_file', { path: await getSettingsPath('settings.json') });
-  };
+  const devTools = useSelector(newSelectors.devTools);
+
+  const dispatch = useDispatch();
+
+  function onToggleDevTools(value: boolean) {
+    dispatch(RootActions.setDevTools(value));
+  }
 
   return (
     <div className={cs.info}>
       <SettingsGroup>
-        <SettingsSubGroup label="Links:">
+        <SettingsSubGroup label="Seelen UI">
           <SettingsOption>
-            <span>
-              Seelen UI <span className={cs.version}>v{EnvConfig.version}</span>:
-            </span>
+            <span>Version:</span>
+            <span className={cs.version}>v{EnvConfig.version}</span>
+          </SettingsOption>
+        </SettingsSubGroup>
+      </SettingsGroup>
+
+      <SettingsGroup>
+        <SettingsSubGroup label="Official Links">
+          <SettingsOption>
+            <span>Github:</span>
             <a href="https://github.com/eythaann/seelen-ui" target="_blank">
               github.com/eythaann/seelen-ui
             </a>
@@ -36,12 +48,8 @@ export function Information() {
 
       <SettingsGroup>
         <SettingsOption>
-          <span>Open Settings file</span>
-          <Button onClick={openSettingsFile}>Open</Button>
-        </SettingsOption>
-        <SettingsOption>
-          <span>Load Custom file (will replace current settings):</span>
-          <Button onClick={LoadCustomConfigFile}>Select File</Button>
+          <span>Development Mode</span>
+          <Switch value={devTools} onChange={onToggleDevTools} />
         </SettingsOption>
       </SettingsGroup>
 
@@ -53,7 +61,7 @@ export function Information() {
           </Button>
         </SettingsOption>
         <SettingsOption>
-          <span>Quit/Close Seelen UI</span>
+          <span>Quit/Close</span>
           <Button type="dashed" danger onClick={() => exit(0)} style={{ width: '50px' }}>
             <Icon iconName="IoClose" />
           </Button>