+++ src/apps/settings/modules/shared/config/infra.ts
@@ -4,6 +4,6 @@ export const EnvConfig = {
   version: process.env.packageVersion,
 };
 
-export async function getSettingsPath(...sub: string[]) {
+export async function resolveDotConfigPath(...sub: string[]) {
   return await path.join(await path.homeDir(), '.config', 'seelen', ...sub);
 }