+++ src/apps/settings/modules/shared/store/app/StateBridge.ts
@@ -62,6 +62,7 @@ export const StateToJsonSettings = (state: RootState): ISettings => {
     'ahkEnabled',
     'fancyToolbar',
     'ahkVariables',
+    'devTools',
   ]);
 };