+++ changelog.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+### enhancements
+- separate `information` and `developer tools` tabs in the settings.
+- add a option to open the install path in explorer.
+
 ## [1.7.0]
 ### features
 - add Network toolbar module.